Fixed issue with enumerable having incorrect values do to point of iteration
The values in the enumerable (which were intended to be static) were changing due to the source object changing between the enumerables instantiation (the values wanted) and its iteration (the values being used). The enumerable is now immediately cast to an array to fix this.

Also addressed https://github.com/QAAFI/DCaPST/pull/4#discussion_r375057175 while fixing.